[FIX] 공 뭐냐 그 투팩터에 넣고, 투팩터 오류 수정하고, 프로필 생긴거 좀 수정
